Přidání a napojení inputProcessoru - přesun zpracování vstupu do samostaného souboru - Funkce sideLength přesunuta z triangle.py do inputProcessor.py - main.py pracuje s listem sides, ve kterém jsou uloženy právě 3 délky stran trojúhelníku - Vylepšené zjišťování odvěsny v main.py při volání triangleOrthogonality pomocí seřazeného pole
